Add support for DuckDB Backend, some fixes to recording (#9)
* Implement DuckDB CUrsor

* Storing current state, not working

* It works but values written are wrong for time series

* Change sequence creation to IF EXISTS

* Fix kernel trace file names in process

* Add duckdb to viz tool

* Add duckdb to path selector

* Improve flow selection

* Fix dport mapping for sock recording

* Udpate

* Fix DataValue parsing from DuckDB Union

* Update README.md

* Move queries to extra file

* Add test db files to gitignore

* Add duckdb dependency to github workflow

* Switch to bundled version

---------

+++ ts-storage/src/sqlite/db.rs
@@ -425,7 +425,7 @@ impl TSDBInterface for SQLiteTSDB {
         let params: &[(_, Value)] = &[
             (":flow_id", id.into()),
             (":name", name.to_string().into()),
-            (":type", ts_type.type_to_int().into()),
+            (":type", (ts_type.type_to_int() as i64).into()),
         ][..];
         
         query.bind::<&[(_, Value)]>(params)?;
